Factorize updates (cli and UI)
Factorize old items and version checks
Factorize updates
Factorize db optimization, crons and plugin deactivation
Plugins are now deactivated at the begin of the update process.
Add few unit tests
Devices list has been updated...
