+++ install/update_to_031.php
@@ -98,4 +98,61 @@ function updateDbTo031() {
              MODIFY `date_fin_garantie` `date` NOT NULL default '0000-00-00'";
 
    echo "<p class='center'>Version 0.3  </p>";
+
+   // Before 0.31
+   if (!TableExists("glpi_config") && !TableExists("glpi_configs")) {
+      $query = "CREATE TABLE `glpi_config` (
+                  `ID` int(11) NOT NULL auto_increment,
+                  `num_of_events` varchar(200) NOT NULL default '',
+                  `jobs_at_login` varchar(200) NOT NULL default '',
+                  `sendexpire` varchar(200) NOT NULL default '',
+                  `cut` varchar(200) NOT NULL default '',
+                  `expire_events` varchar(200) NOT NULL default '',
+                  `list_limit` varchar(200) NOT NULL default '',
+                  `version` varchar(200) NOT NULL default '',
+                  `logotxt` varchar(200) NOT NULL default '',
+                  `root_doc` varchar(200) NOT NULL default '',
+                  `event_loglevel` varchar(200) NOT NULL default '',
+                  `mailing` varchar(200) NOT NULL default '',
+                  `imap_auth_server` varchar(200) NOT NULL default '',
+                  `imap_host` varchar(200) NOT NULL default '',
+                  `ldap_host` varchar(200) NOT NULL default '',
+                  `ldap_basedn` varchar(200) NOT NULL default '',
+                  `ldap_rootdn` varchar(200) NOT NULL default '',
+                  `ldap_pass` varchar(200) NOT NULL default '',
+                  `admin_email` varchar(200) NOT NULL default '',
+                  `mailing_signature` varchar(200) NOT NULL default '',
+                  `mailing_new_admin` varchar(200) NOT NULL default '',
+                  `mailing_followup_admin` varchar(200) NOT NULL default '',
+                  `mailing_finish_admin` varchar(200) NOT NULL default '',
+                  `mailing_new_all_admin` varchar(200) NOT NULL default '',
+                  `mailing_followup_all_admin` varchar(200) NOT NULL default '',
+                  `mailing_finish_all_admin` varchar(200) NOT NULL default '',
+                  `mailing_new_all_normal` varchar(200) NOT NULL default '',
+                  `mailing_followup_all_normal` varchar(200) NOT NULL default '',
+                  `mailing_finish_all_normal` varchar(200) NOT NULL default '',
+                  `mailing_new_attrib` varchar(200) NOT NULL default '',
+                  `mailing_followup_attrib` varchar(200) NOT NULL default '',
+                  `mailing_finish_attrib` varchar(200) NOT NULL default '',
+                  `mailing_new_user` varchar(200) NOT NULL default '',
+                  `mailing_followup_user` varchar(200) NOT NULL default '',
+                  `mailing_finish_user` varchar(200) NOT NULL default '',
+                  `ldap_field_name` varchar(200) NOT NULL default '',
+                  `ldap_field_email` varchar(200) NOT NULL default '',
+                  `ldap_field_location` varchar(200) NOT NULL default '',
+                  `ldap_field_realname` varchar(200) NOT NULL default '',
+                  `ldap_field_phone` varchar(200) NOT NULL default '',
+               PRIMARY KEY (`ID`)
+               ) TYPE=MyISAM AUTO_INCREMENT=2 ";
+      $DB->queryOrDie($query);
+
+      $query = "INSERT INTO `glpi_config`
+               VALUES (1, '10', '1', '1', '80', '30', '15', ' 0.31', 'GLPI powered by indepnet',
+                        '/glpi', '5', '0', '', '', '', '', '', '', 'admsys@xxxxx.fr', 'SIGNATURE',
+                        '1', '1', '1', '1', '0', '0', '0', '0', '0', '0', '0', '0','1', '1', '1',
+                        'uid', 'mail', 'physicaldeliveryofficename', 'cn', 'telephonenumber')";
+      $DB->queryOrDie($query);
+
+      echo "<p class='center'>Version > 0.31  </p>";
+   }
 }
